requirements.txt
asgiref==3.8.1
boto3==1.35.11
botocore==1.35.11
build==1.2.1
chardet==5.2.0
click==8.1.7
cssselect2==0.7.0
dj-database-url==2.2.0
Django==5.0.9
django-classy-tags==4.1.0
django-cms==4.1.3
django-debug-toolbar==4.4.6
django-entangled==0.5.4
django-filer==3.2.0
django-formtools==2.5.1
django-fsm==2.8.2
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-storage-url==0.12.1
django-storages==1.14.4
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==3.0.0
djangocms-frontend==1.3.3
djangocms-text==0.2.3
djangocms-versioning==2.0.2
easy-thumbnails==2.9
furl==2.1.3
jmespath==1.0.1
lxml==5.3.0
nh3==0.2.18
orderedmultidict==1.0.1
packaging==24.1
pillow==10.4.0
pip-tools==7.4.1
psycopg2==2.9.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
reportlab==4.2.2
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==2.2.2
uWSGI==2.0.26
webencodings==0.5.1
whitenoise==6.7.0
